# frontend/src/shared/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateModelConnectionRequest": {
        "additionalProperties": false,
        "properties": {
          "auth_mode": {
            "enum": [
              "chatgpt_subscription",
              "api_key",
              "local",
              "none"
            ],
            "title": "Auth Mode",
            "type": "string"
          },
          "base_url": {
            "anyOf": [
              {
                "maxLength": 2048,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "credential": {
            "anyOf": [
              {
                "maxLength": 16384,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credential"
          },
          "display_name": {
            "anyOf": [
              {
                "maxLength": 160,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "model_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Model Id",
            "type": "string"
          },
          "provider_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Provider Id",
            "type": "string"
          }
        },
        "required": [
          "provider_id",
          "model_id",
          "auth_mode"
        ],
        "title": "CreateModelConnectionRequest",
        "type": "object"
      },
...
      "EventListResponse": {
        "additionalProperties": false,
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/ToxAgentEvent"
            },
            "title": "Events",
            "type": "array"
          },
          "latest_sequence": {
            "title": "Latest Sequence",
            "type": "integer"
          }
        },
        "required": [
          "events",
          "count",
          "latest_sequence"
        ],
        "title": "EventListResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ModelConnection": {
        "additionalProperties": true,
        "properties": {
          "auth_mode": {
            "title": "Auth Mode",
            "type": "string"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "capabilities": {
            "$ref": "#/components/schemas/ModelConnectionCapabilities"
          },
          "connection_id": {
            "title": "Connection Id",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "has_credential": {
            "title": "Has Credential",
            "type": "boolean"
          },
          "model_id": {
            "title": "Model Id",
            "type": "string"
          },
          "provider_id": {
            "title": "Provider Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "untested",
              "ready",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "connection_id",
          "provider_id",
          "model_id",
          "display_name",
          "auth_mode",
          "base_url",
          "has_credential",
          "capabilities",
          "status"
        ],
        "title": "ModelConnection",
        "type": "object"
      },
      "ModelConnectionCapabilities": {
        "additionalProperties": true,
        "properties": {
          "context_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Size"
          },
          "streaming": {
            "title": "Streaming",
            "type": "boolean"
          },
          "structured_output": {
            "title": "Structured Output",
            "type": "boolean"
          },
          "tool_calls": {
            "title": "Tool Calls",
            "type": "boolean"
          }
        },
        "required": [
          "streaming",
          "tool_calls",
          "structured_output",
          "context_size"
        ],
        "title": "ModelConnectionCapabilities",
        "type": "object"
      },
      "ModelConnectionList": {
        "additionalProperties": false,
        "properties": {
          "connections": {
            "items": {
              "$ref": "#/components/schemas/ModelConnection"
            },
            "title": "Connections",
            "type": "array"
          }
        },
        "required": [
          "connections"
        ],
        "title": "ModelConnectionList",
        "type": "object"
      },
...
      "ToxAgentEvent": {
        "additionalProperties": false,
        "properties": {
          "entity_id": {
            "title": "Entity Id",
            "type": "string"
          },
          "entity_type": {
            "title": "Entity Type",
            "type": "string"
          },
          "entity_version": {
            "title": "Entity Version",
            "type": "integer"
          },
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "occurred_at": {
            "title": "Occurred At",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "sequence": {
            "title": "Sequence",
            "type": "integer"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "event_id",
          "session_id",
          "sequence",
          "type",
          "entity_type",
          "entity_id",
          "entity_version",
          "run_id",
          "occurred_at",
          "payload"
        ],
        "title": "ToxAgentEvent",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "ToxAgent control plane.\n\nEvidence and decision support over the ToxPred predictor. Every number in an\naccepted answer resolves to a stored observation and a field path; hERG, Tox21\nand ClinTox stay three separate measurements; and no schema here can carry an\naggregate toxicity or safety verdict.\n",
    "title": "ToxAgent control plane",
    "version": "0.1.0.dev0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/model-connections": {
      "get": {
        "operationId": "list_model_connections_v1_model_connections_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelConnectionList"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Model Connections",
        "tags": [
          "toxagent"
        ]
      },
      "post": {
        "operationId": "create_model_connection_v1_model_connections_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateModelConnectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelConnection"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Model Connection",
        "tags": [
          "toxagent"
        ]
      }
    },
    "/v1/model-connections/{connection_id}": {
      "delete": {
        "operationId": "delete_model_connection_v1_model_connections__connection_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "connection_id",
            "required": true,
            "schema": {
              "title": "Connection Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Model Connection",
        "tags": [
          "toxagent"
        ]
      },
      "get": {
        "operationId": "get_model_connection_v1_model_connections__connection_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "connection_id",
            "required": true,
            "schema": {
              "title": "Connection Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelConnection"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Model Connection",
        "tags": [
          "toxagent"
        ]
      }
    },
    "/v1/model-connections/{connection_id}:test": {
      "post": {
        "operationId": "test_model_connection_v1_model_connections__connection_id__test_post",
        "parameters": [
          {
            "in": "path",
            "name": "connection_id",
            "required": true,
            "schema": {
              "title": "Connection Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelConnection"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Test Model Connection",
        "tags": [
          "toxagent"
        ]
      }
    },
...
    "/v1/sessions/{session_id}/events": {
      "get": {
        "operationId": "stream_events_v1_sessions__session_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after_sequence",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "After Sequence",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Events",
        "tags": [
          "toxagent"
        ]
      }
    },
    "/v1/sessions/{session_id}/events:list": {
      "get": {
        "description": "A non-streaming read of the same outbox the SSE feed serves.\n\nThe stream never terminates, so it is the wrong tool for \"replay\neverything that happened in this one run\" (a Run Inspector opened after\nthe fact, or a page that reconnected and needs to fill a gap). The outbox\nrow is retained forever — nothing here is a delivery guarantee beyond what\n``/events`` already gives; this just lets a client stop listening.",
        "operationId": "list_events_v1_sessions__session_id__events_list_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after_sequence",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "After Sequence",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "run_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Events",
        "tags": [
          "toxagent"
        ]
      }
    },
...
  }
}